Moves the Prelude into this project (previously, it was copied from CMC).  This is in preparation for creating a proper Curry library.

+++ Makefile
@@ -79,12 +79,6 @@ $(PREFIX)/lib/curry : | $(PREFIX)/lib
 	mkdir -p $(PREFIX)/lib/curry
 $(PREFIX)/lib/curry/.curry : | $(PREFIX)/lib/curry
 	mkdir -p $(PREFIX)/lib/curry/.curry
-$(PREFIX)/lib/curry/Prelude.curry : $(CMC_HOME)/runtime/lib/Prelude.curry | $(PREFIX)/lib/curry
-	cp $< $@
-$(PREFIX)/lib/curry/.curry/Prelude.%    : $(CMC_HOME)/runtime/lib/.curry/Prelude.% \
-                                          $(PREFIX)/lib/curry/Prelude.curry        \
-                                        | $(PREFIX)/lib/curry/.curry
-	cp $< $@
 
 $(ROOT_DIR)/install:
 	ln -s $(PREFIX) $@
@@ -97,15 +91,31 @@ install: $(PREFIX)/.bin/coverage                   \
          $(PREFIX)/bin/python                      \
          $(PREFIX)/bin/icy                         \
          $(PREFIX)/bin/curryexec                   \
-         $(PREFIX)/lib/curry/Prelude.curry         \
-         $(PREFIX)/lib/curry/.curry/Prelude.fcy    \
-         $(PREFIX)/lib/curry/.curry/Prelude.fint   \
-         $(PREFIX)/lib/curry/.curry/Prelude.icur   \
-         $(PREFIX)/lib/curry/.curry/Prelude.icurry \
-         $(PREFIX)/lib/curry/.curry/Prelude.json   \
-         $(PREFIX)/lib/curry/.curry/Prelude.read   \
+				 install_currylib
   ####
 	@echo "\n****** Sprite is installed under $(PREFIX) ******\n"
+
+# The the Curry library, copy files from currylib/.
+CURRYLIB_CURRY_FILES = $(addprefix                    \
+    $(PREFIX)/lib/curry/,                             \
+    $(notdir $(wildcard $(ROOT_DIR)currylib/*.curry)) \
+  )
+CURRYLIB_DOTCURRY_FILES = $(addprefix                  \
+    $(PREFIX)/lib/curry/.curry/,                       \
+    $(notdir $(wildcard $(ROOT_DIR)currylib/.curry/*)) \
+  )
+
+.PHONY: install_currylib
+install_currylib: $(CURRYLIB_CURRY_FILES) $(CURRYLIB_DOTCURRY_FILES) \
+                | $(PREFIX)/lib/curry $(PREFIX)/lib/curry/.curry
+
+$(PREFIX)/lib/curry/%.curry: $(ROOT_DIR)currylib/%.curry \
+    | $(PREFIX)/lib/curry
+	cp $< $@
+
+$(PREFIX)/lib/curry/.curry/%: $(ROOT_DIR)currylib/.curry/% \
+    | $(PREFIX)/lib/curry/.curry
+	cp $< $@
 
 ifneq ($(PREFIX),python)
 install: $(ROOT_DIR)/install
@@ -120,12 +130,15 @@ uninstall:
 	-rm -f $(PREFIX)/bin/python
 	-rm -f $(PREFIX)/bin/icy
 	-rm -f $(PREFIX)/bin/curryexec
-	-rm -f $(PREFIX)/lib/curry/Prelude.curry
+	-rm -f $(PREFIX)/lib/curry/*.curry
 	-rm -rf $(PREFIX)/lib/curry/.curry/*
-	-rmdir $(PREFIX)/bin
-	-rmdir $(PREFIX)/.bin
-	-rmdir $(PREFIX)/lib/curry/.curry
-	-rmdir $(PREFIX)/lib/curry
-	-rmdir $(PREFIX)/lib
+	-if [ -d $(PREFIX)/bin ];              then rmdir $(PREFIX)/bin;              fi
+	-if [ -d $(PREFIX)/.bin ];             then rmdir $(PREFIX)/.bin;             fi
+	-if [ -d $(PREFIX)/lib/curry/.curry ]; then rmdir $(PREFIX)/lib/curry/.curry; fi
+	-if [ -d $(PREFIX)/lib/curry ];        then rmdir $(PREFIX)/lib/curry;        fi
+	-if [ -d $(PREFIX)/lib ];              then rmdir $(PREFIX)/lib;              fi
+ifeq ($(PREFIX),$(ROOT_DIR)/install)
+	echo YES
 	-rmdir $(ROOT_DIR)/install
+endif
 endif
